Moved the base specification classes so that they are exported and can be used by other projects.
