+++ build.gradle
@@ -56,7 +56,7 @@ dependencies {
 
     implementation 'org.codehaus.groovy.modules.http-builder:http-builder:0.5.0'
 
-    testImplementation "org.spockframework:spock-core:1.2-groovy-2.5"
+    implementation "org.spockframework:spock-core:1.2-groovy-2.5"
     testImplementation 'junit:junit:4.12'
     testImplementation group: 'cglib', name: 'cglib-nodep', version: '3.3.0'
 }
