+++ src/test/groovy/me/biocomp/hubitat_ci/integration/DimmerMinimumsIntegrationTest.groovy
@@ -1,6 +1,6 @@
 package me.biocomp.hubitat_ci.integration
 
-import me.biocomp.hubitat_ci.integration.IntegrationAppSpecification
+import me.biocomp.hubitat_ci.util.integration.IntegrationAppSpecification
 import me.biocomp.hubitat_ci.util.device_fixtures.DimmerFixtureFactory
 
 import spock.lang.Specification
@@ -21,7 +21,7 @@ class DimmerMinimumsIntegrationTest extends IntegrationAppSpecification {
         dimmerFixture1 = DimmerFixtureFactory.create('d1')
         dimmerFixture2 = DimmerFixtureFactory.create('d2')
 
-        super.initializeEnvironment("Scripts/DimmerMinimums.groovy", [dimmers: [dimmerFixture1, dimmerFixture2], minimumLevel: 5, enableLogging: true])
+        super.initializeEnvironment("Scripts/DimmerMinimums.groovy", [], [dimmers: [dimmerFixture1, dimmerFixture2], minimumLevel: 5, enableLogging: true])
     }
 
     void "App initialize subscribes to events"() {
